--- src/main/java/gift/Wishlist/WishList.java
@@ -1,60 +1,0 @@
-package gift.Wishlist;
-
-import gift.member.Member;
-import gift.product.Product;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-
-@Entity
-@Table(name = "wishlist")
-public class WishList {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_Id")
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_Id")
-    private Product product;
-
-    @Column(nullable = false)
-    private Integer quantity;
-
-    public WishList(Member member, Product product, Integer quantity){
-        this.member = member;
-        this.product = product;
-        this.quantity = quantity;
-    }
-
-    public void updateQuantity(int quantity){
-        this.quantity += quantity;
-    }
-
-    protected WishList() {}
-
-    public Long getId() {
-        return id;
-    }
-
-    public Member member() {
-        return member;
-    }
-
-    public Product getProduct(){
-        return product;
-    }
-
-    public Integer getQuantity(){
-        return quantity;
-    }
-}
