--- src/main/java/gift/Wishlist/dto/WishResponseDto.java
@@ -1,11 +1,0 @@
-package gift.Wishlist.dto;
-
-public record WishResponseDto(
-        Long wishListId,
-        String productName,
-        String productImage,
-        Integer quantity,
-        Integer price
-) {
-
-}
