--- src/main/java/gift/Wishlist/exception/WishNotFoundException.java
@@ -1,11 +1,0 @@
-package gift.Wishlist.exception;
-
-import gift.exception.ErrorCode;
-import gift.exception.MyException;
-
-public class WishNotFoundException extends MyException {
-
-    public WishNotFoundException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
