--- src/main/java/gift/option/dto/OptionResponseDto.java
@@ -1,8 +1,0 @@
-package gift.option.dto;
-
-public record OptionResponseDto(
-        Long optionId,
-        String optionName,
-        Integer quantity,
-        Integer price
-){ }
