--- src/main/java/gift/option/exception/OptionNotFound.java
@@ -1,11 +1,0 @@
-package gift.option.exception;
-
-import gift.exception.ErrorCode;
-import gift.exception.MyException;
-
-public class OptionNotFound extends MyException {
-
-  public OptionNotFound(ErrorCode errorCode) {
-    super(errorCode);
-  }
-}
