+++ src/main/java/gift/product/controller/api/ProductController.java
@@ -1,7 +1,7 @@
 package gift.product.controller.api;
 
-import gift.option.dto.OptionRequestDto;
-import gift.option.service.OptionService;
+import gift.productoption.dto.OptionRequestDto;
+import gift.productoption.service.OptionService;
 import gift.product.dto.ProductOptionRequestDto;
 import gift.product.Product;
 import gift.exception.ErrorCode;
