+++ src/main/java/gift/productoption/controller/api/OptionController.java
@@ -1,0 +1,60 @@
+package gift.productoption.controller.api;
+
+import gift.productoption.dto.OptionRequestDto;
+import gift.productoption.dto.OptionResponseDto;
+import gift.productoption.service.OptionService;
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api")
+public class OptionController {
+
+    @Autowired OptionService optionService;
+
+    @PostMapping("/options/{productId}")
+    public ResponseEntity<OptionResponseDto> createOption(
+            @PathVariable Long productId,
+            @RequestBody OptionRequestDto optionRequestDto
+    ){
+        OptionResponseDto responseDto = optionService.createOptionByProductId(productId, optionRequestDto);
+        return new ResponseEntity<>(responseDto, HttpStatus.CREATED);
+    }
+
+    //상품아이디를 통해 특정 상품의 옵션을 모두 조회
+    @GetMapping("/products/{productId}/options")
+    public ResponseEntity<List<OptionResponseDto>> getAllProductOptions(@PathVariable Long productId){
+        return ResponseEntity.ok(optionService.getOptionByProduct(productId));
+    }
+
+    //optionId를 통해 특정 옵션을 조회
+    @GetMapping("/options/{optionId}")
+    public ResponseEntity<OptionResponseDto> getOption(@PathVariable Long optionId){
+        return ResponseEntity.ok(optionService.findOne(optionId));
+    }
+
+    @PutMapping("/options/{optionId}")
+    public ResponseEntity<OptionResponseDto> editOption(
+            @PathVariable Long optionId,
+            @RequestBody OptionRequestDto optionRequestDto
+    ){
+        return ResponseEntity.ok(optionService.updateOption(optionId, optionRequestDto));
+    }
+
+    @DeleteMapping("/options/{optionId}")
+    public ResponseEntity<Void> removeOption(@PathVariable Long optionId){
+        optionService.removeOtion(optionId);
+        return ResponseEntity.noContent().build();
+    }
+
+}
