+++ src/main/java/gift/productoption/dto/OptionResponseDto.java
@@ -1,0 +1,8 @@
+package gift.productoption.dto;
+
+public record OptionResponseDto(
+        Long id,
+        String optionName,
+        Integer quantity,
+        Integer price
+){ }
