+++ src/main/java/gift/productoption/exception/OptionPriceError.java
@@ -1,0 +1,11 @@
+package gift.productoption.exception;
+
+import gift.exception.ErrorCode;
+import gift.exception.MyException;
+
+public class OptionPriceError extends MyException {
+
+    public OptionPriceError(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
