+++ src/main/java/gift/productoption/exception/UnavailableOptionName.java
@@ -1,0 +1,11 @@
+package gift.productoption.exception;
+
+import gift.exception.ErrorCode;
+import gift.exception.MyException;
+
+public class UnavailableOptionName extends MyException {
+
+  public UnavailableOptionName(ErrorCode errorCode) {
+    super(errorCode);
+  }
+}
