+++ src/main/java/gift/wishlist/dto/WishResponseDto.java
@@ -1,0 +1,13 @@
+package gift.wishlist.dto;
+
+public record WishResponseDto(
+        Long wishListId,
+        Long optionId,
+        String productName,
+        String optionName,
+        String productImage,
+        Integer quantity,
+        Integer price
+) {
+
+}
