+++ src/main/java/gift/wishlist/exception/WishNotFoundException.java
@@ -1,0 +1,11 @@
+package gift.wishlist.exception;
+
+import gift.exception.ErrorCode;
+import gift.exception.MyException;
+
+public class WishNotFoundException extends MyException {
+
+    public WishNotFoundException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
