+++ src/main/resources/application.properties
@@ -9,7 +9,7 @@ spring.datasource.password=
 
 spring.jpa.hibernate.ddl-auto=none
 spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.show-sql=true
+#spring.jpa.show-sql=true
 
 spring.jpa.defer-datasource-initialization=true
 spring.sql.init.mode=always
