+++ src/main/resources/schema.sql
@@ -17,7 +17,7 @@ create table member
     unique (email)
 );
 
-create table option
+create table product_option
 (
     id         bigint auto_increment,
     name       varchar(255) not null,
@@ -32,9 +32,9 @@ create table wishlist
 (
     id         bigint auto_increment,
     member_id  bigint not null,
-    product_id bigint not null,
+    option_id bigint not null,
     quantity   int    not null,
     primary key (id),
-    foreign key (product_id) references product (id),
+    foreign key (option_id) references product (id),
     foreign key (member_id) references member (id)
 );
