+++ src/main/resources/templates/yjshop/admin/product/modifyform.html
@@ -29,8 +29,40 @@
       <div class="form-buttons">
         <input type="submit" value="상품 수정하기">
       </div>
-
     </form>
   </div>
+
+  <table>
+    <th>옵션</th>
+    <th>수량</th>
+    <th>가격</th>
+    <tr th:each="option: ${options}">
+      <form th:action="@{/view/admin/options/edit/{id}(id=${option.id})}" method="post" class="admin-form">
+        <div>
+          <label><strong>옵션</strong></label>
+          <input type="text" name="name" placeholder="상품명 입력" th:value="${option.optionName}">
+        </div>
+
+        <div>
+          <label><strong>가격</strong></label>
+          <input type="text" name="name" placeholder="상품명 입력" th:value="${option.price}">
+        </div>
+
+        <div>
+          <label><strong>수량</strong></label>
+          <input type="text" name="name" placeholder="상품명 입력" th:value="${option.quantity()}">
+        </div>
+
+        <input type="hidden" name="productId" th:value="${option.id}">
+        <button type="submit">수정</button>
+
+        <form th:action="@{/view/admin/options/remove/{id}(id=${option.id})}" method="post">
+            <button type="submit">삭제</button>
+        </form>
+
+      </form>
+    </tr>
+  </table>
+
 </div>
 </html>
