--- src/main/resources/templates/yjshop/admin/product/productoptions.html
@@ -1,74 +1,0 @@
-<html layout:decorate="~{yjshop/user/layout}">
-<div layout:fragment="content">
-  <table>
-
-    <style>
-      /* ✅ Actions 칸 내 form 요소들을 수평으로 정렬 */
-      td form {
-        display: inline-flex;
-        align-items: center;
-        gap: 6px;
-        margin: 5px;
-      }
-
-      /* ✅ 수량 입력창 크기 통일 */
-      td input[type="number"] {
-        width: 70px;
-        height: 32px;
-        padding: 4px;
-        font-size: 14px;
-        box-sizing: border-box;
-      }
-
-      /* ✅ 버튼 스타일 통일 */
-      td button,
-      td input[type="submit"] {
-        height: 32px;
-        padding: 0 10px;
-        font-size: 14px;
-        background-color: #90ee90;
-        border: 1px solid #ccc;
-        font-weight: bold;
-        cursor: pointer;
-        box-sizing: border-box;
-      }
-    </style>
-
-    <thead>
-    <tr>
-      <th>상품명</th>
-      <th>가격</th>
-      <th>이미지</th>
-      <th>Actions</th>
-    </tr>
-    </thead>
-    <tbody>
-    <tr>
-      <td th:text="${product.name}"></td>
-      <td th:text="${product.price}"></td>
-      <td><img th:src="${product.imageUrl}" width="100px" alt="상품 이미지"></td>
-
-    </tr>
-    </tbody>
-  </table>
-
-  <table>
-    <tr th:each="option: ${options}">
-      <td th:text="${option.optionName}"></td>
-      <td th:text="${option.price}"></td>
-      <td th:text="${option.quantity}"></td>
-      <td>
-        <form th:action="@{/view/my/wishlist}" method="post">
-          <input type="hidden" name="productId" th:value="${product.id}">
-          <input type="number" name="quantity" placeholder="수량">
-          <button type="submit">수정</button>
-        </form>
-        <form th:action="@{/view/my/payment}" method="get">
-          <button type="submit">삭제</button>
-        </form>
-      </td>
-    </tr>
-  </table>
-
-</div>
-</html>
