+++ src/main/resources/templates/yjshop/user/home.html
@@ -35,7 +35,7 @@
       <th>상품명</th>
       <th>가격</th>
       <th>이미지</th>
-      <th>구매하기 / 장바구니 담기</th>
+<!--      <th>구매하기 / 장바구니 담기</th>-->
     </tr>
     </thead>
     <tbody>
@@ -48,19 +48,19 @@
         </a>
       </td>
 
-      <td>
-        <form th:action="@{/view/my/payment}" method="get">
-          <input type="number" name="quantity" placeholder="수량">
-          <input type="hidden" name="productId" th:value="${product.id}">
-          <button type="submit">상품 구매</button>
-        </form>
+<!--      <td>-->
+<!--        <form th:action="@{/view/my/payment}" method="get">-->
+<!--          <input type="number" name="quantity" placeholder="수량">-->
+<!--          <input type="hidden" name="productId" th:value="${product.id}">-->
+<!--          <button type="submit">상품 구매</button>-->
+<!--        </form>-->
 
-        <form th:action="@{/view/my/wishlist}" method="post">
-          <input type="hidden" name="productId" th:value="${product.id}">
-          <input type="number" name="quantity" placeholder="수량">
-          <button type="submit">장바구니 담기</button>
-        </form>
-      </td>
+<!--        <form th:action="@{/view/my/wishlist}" method="post">-->
+<!--          <input type="hidden" name="productId" th:value="${product.id}">-->
+<!--          <input type="number" name="quantity" placeholder="수량">-->
+<!--          <button type="submit">장바구니 담기</button>-->
+<!--        </form>-->
+<!--      </td>-->
     </tr>
     </tbody>
   </table>
