+++ src/main/resources/templates/yjshop/user/wishlist.html
@@ -30,41 +30,49 @@
 
     </style>
 
-  <thead>
-  <tr>
-    <th>상품명</th>
-    <th>상품이미지</th>
-    <th>수량</th>
-    <th>가격</th>
-    <th>장바구니 수정</th>
-  </tr>
-  </thead>
+    <thead>
+    <tr>
+      <th>상품명</th>
+      <th>상품 옵션</th>
+      <th>상품이미지</th>
+      <th>수량</th>
+      <th>가격</th>
+      <th>장바구니 수정</th>
+    </tr>
+    </thead>
 
-  <tbody>
-  <tr th:each="responseDto: ${wishlist}">
-    <td th:text="${responseDto.productName}"></td>
-    <td><img th:src="@{${responseDto.productImage}}" width="100px" alt="상품 이미지"/></td>
-    <td th:text="${responseDto.quantity}"></td>
-    <td th:text="${responseDto.price}"></td>
+    <tbody>
+    <tr th:each="responseDto: ${wishlist}">
+      <td th:text="${responseDto.productName}"></td>
+      <td th:text="${responseDto.optionName}"></td>
+      <td><img th:src="@{${responseDto.productImage}}" width="100px" alt="상품 이미지"/></td>
+      <td th:text="${responseDto.quantity}"></td>
+      <td th:text="${responseDto.price}"></td>
 
-    <td>
-      <form th:action="@{/view/my/wishlist/add/{wishListId}(wishListId=${responseDto.wishListId})}" method="post">
-        <button type="submit">+</button>
-      </form>
-      <form th:action="@{/view/my/wishlist/subtract/{wishListId}(wishListId=${responseDto.wishListId})}" method="post">
-        <button type="submit">-</button>
-      </form>
-      <form th:action="@{/view/my/payment}" method="get">
-        <button type="submit">바로 구매하기</button>
-      </form>
-      <form th:action="@{/view/my/wishlist/delete/{wishListId}(wishListId=${responseDto.wishListId})}" method="post">
-        <button type="submit">비우기</button>
-      </form>
-    </td>
+      <td>
+        <form
+            th:action="@{/view/my/wishlist/add/{wishListId}(wishListId=${responseDto.wishListId})}"
+            method="post">
+          <button type="submit">+</button>
+        </form>
+        <form
+            th:action="@{/view/my/wishlist/subtract/{wishListId}(wishListId=${responseDto.wishListId})}"
+            method="post">
+          <button type="submit">-</button>
+        </form>
+        <form th:action="@{/view/my/payment}" method="get">
+          <button type="submit">바로 구매하기</button>
+        </form>
+        <form
+            th:action="@{/view/my/wishlist/delete/{wishListId}(wishListId=${responseDto.wishListId})}"
+            method="post">
+          <button type="submit">비우기</button>
+        </form>
+      </td>
 
-  </tr>
-  </tbody>
-</table>
+    </tr>
+    </tbody>
+  </table>
 
   <div th:if="${!wishlist.empty}">
     <ul class="pagination justify-content-center">
@@ -106,6 +114,6 @@
     </ul>
   </div>
 
-</body>
+  </body>
 </div>
 </html>
